changed band type to .n and adjusted c* api

+++ index/storage_cass.go
@@ -53,7 +53,7 @@ func (c *CassandraStorage) Fetch(bitmap_id uint64, db string, frame string, slic
 		chunk_key, block int64
 		block_index      uint32
 		s8               uint8
-		filter           uint64
+		filter           int
 	)
 	log.Println("FETCHING ", bitmap_id, db, slice)
 
@@ -77,7 +77,7 @@ func (c *CassandraStorage) Fetch(bitmap_id uint64, db string, frame string, slic
 
 	}
 	bitmap.SetCount(uint64(count))
-	return bitmap, filter
+	return bitmap, uint64(filter)
 }
 
 func (c *CassandraStorage) Store(id int64, db string, frame string, slice int, filter uint64, bitmap *Bitmap) error {
@@ -103,7 +103,7 @@ func (c *CassandraStorage) Store(id int64, db string, frame string, slice int, f
 func (c *CassandraStorage) StoreBlock(id int64, db string, frame string, slice int, filter uint64, chunk int64, block_index int32, block int64) error {
 
 	if err := c.db.Query(`
-        INSERT INTO bitmap ( bitmap_id, db, frame, slice , filter, ChunkKey, BlockIndex, block) VALUES (?,?,?,?,?,?,?,?);`, id, db, frame, slice, filter, chunk, block_index, block).Exec(); err != nil {
+        INSERT INTO bitmap ( bitmap_id, db, frame, slice , filter, ChunkKey, BlockIndex, block) VALUES (?,?,?,?,?,?,?,?);`, id, db, frame, slice, int(filter), chunk, block_index, block).Exec(); err != nil {
 		log.Println(err)
 		log.Println("INSERT ", id, chunk, block_index)
 		return err
